requirements.txt
pandas
numpy
pyaml
sys
os
scikit-learn
-e
